+++ R/vbStarts.R
@@ -10,8 +10,6 @@
 #' 
 #' Starting values for the L1 and L3 parameters in the Schnute paramaterization and the L1, L2, and L3 parameters in the Francis parameterization may be found in two ways.  If \code{methEV="poly"}, then the starting values are the predicted length-at-age from a second-degree polynomial fit to the mean lengths-at-age data.  If \code{methEV="means"} then the observed sample means at the corresponding ages are used.  In the case where one of the supplied ages is fractional, then the value returned will be linearly interpolated between the mean lengths of the two closest ages.  The ages to be used for L1 and L3 in the Schnute and Francis parameterizations are supplied as a numeric vector of length 2 in \code{ages2use=}.  If \code{ages2use=NULL} then the minimum and maximum observed ages will be used.  In the Francis method, L2 will correspond to the age half-way between the two ages in \code{ages2use=}.  A warning will be given if L2<L1 for the Schnute method or if L2<L1 or L3<L2 for the Francis method.
 #' 
-#' Starting values can also be obtained by plotting the data with the model superimposed but tied to slider bars that allow the parameters to be interactively changed.  One can change the parameters until a reasonable fit is observed and then use those valeus as starting values.  Such a plot is constructed by including \code{dynamicPlot=TRUE}.  The initial parameters for the slider bars are the starting values constructed as described above.  It should be noted that the dynamic plot may show an error of \dQuote{[tcl] can't get device image}, but the plot will correctly update if the slider bar is adjusted.
-#' 
 #' @param formula A formula of the form \code{len~age}.
 #' @param data A data frame that contains the variables in \code{formula}.
 #' @param type A string that indicates the parameterization of the von Bertalanffy model.
@@ -23,7 +21,8 @@
 #' @param lwd.mdl A line width for the model when \code{plot=TRUE}.
 #' @param lty.mdl A line type for the model when \code{plot=TRUE}.
 #' @param cex.main A character expansion value for the main title when \code{plot=TRUE}.
-#' @param dynamicPlot A logical that indicates whether a plot with dynamically linked slider bars should be constructed for finding starting values.
+#' @param col.main A color for the main title when \code{plot=TRUE}.
+#' @param dynamicPlot DEPRECATED.
 #' @param \dots Further arguments passed to the methods.
 #' 
 #' @return A list that contains reasonable starting values.  Note that the parameters will be listed in the same order and with the same names as listed in \code{\link{vbFuns}}.
@@ -73,17 +72,6 @@
 #' vbStarts(tl~age,data=SpotVA1,type="Somers",plot=TRUE)
 #' vbStarts(tl~age,data=SpotVA1,type="Somers2",plot=TRUE)
 #' 
-#' ## Dynamic Plots Method -- ONLY RUN IN INTERACTIVE MODE
-#' if (interactive()) {
-#'   vbStarts(tl~age,data=SpotVA1,dynamicPlot=TRUE)
-#'   vbStarts(tl~age,data=SpotVA1,type="original",dynamicPlot=TRUE)
-#'   vbStarts(tl~age,data=SpotVA1,type="GQ",dynamicPlot=TRUE)
-#'   vbStarts(tl~age,data=SpotVA1,type="Mooij",dynamicPlot=TRUE)
-#'   vbStarts(tl~age,data=SpotVA1,type="Weisberg",dynamicPlot=TRUE)
-#'   vbStarts(tl~age,data=SpotVA1,type="Francis",ages2use=c(0,5),dynamicPlot=TRUE)
-#'   vbStarts(tl~age,data=SpotVA1,type="Schnute",ages2use=c(0,5),dynamicPlot=TRUE)
-#' } 
-#' 
 #' ## See examples in vbFuns() for use of vbStarts() when fitting Von B models
 #' 
 #' @export vbStarts
@@ -93,8 +81,8 @@ vbStarts <- function(formula,data=NULL,
                             "GQ","GallucciQuinn","Mooij","Weisberg",
                             "Schnute","Francis","Somers","Somers2"),
                      ages2use=NULL,methEV=c("means","poly"),meth0=c("yngAge","poly"),
-                     plot=FALSE,col.mdl="gray70",lwd.mdl=3,lty.mdl=1,cex.main=0.75,
-                     dynamicPlot=FALSE,...) {
+                     plot=FALSE,col.mdl="gray70",lwd.mdl=3,lty.mdl=1,
+                     cex.main=0.9,col.main="red",dynamicPlot=FALSE,...) {
   ## some checks of arguments
   type <- match.arg(type)
   methEV <- match.arg(methEV)
@@ -126,16 +114,12 @@ vbStarts <- function(formula,data=NULL,
     Somers=                  { sv <- iVBStarts.Somers(age,len,type,meth0) },
     Somers2=                 { sv <- iVBStarts.Somers2(age,len,type,meth0) }
   ) # end 'type' switch
+  ## make the static plot if asked for
+  if (plot) iVBStartsPlot(age,len,type,sv,ages2use,col.mdl,lwd.mdl,lty.mdl,cex.main,col.main)
   ## Check if user wants to choose starting values from an interactive plot
-  if (dynamicPlot) {
-    if (!requireNamespace("relax")) stop("'vbStarts' requires the 'relax' package to be installed to construct the dynamic plot.",call.=FALSE)
-    else iVBStartsDynPlot(age,len,type,sv,ages2use)
-  } else {
-    # make the static plot if asked for
-    if (plot) iVBStartsPlot(age,len,type,sv,ages2use,col.mdl,lwd.mdl,lty.mdl,cex.main)
-    ## return starting values list
-    sv
-  }
+  if (dynamicPlot) warning("The 'dynamicPlot' functionality has been moved to 'vbStartsDP' in the 'FSASims' package.",call.=FALSE)
+  ## return starting values list
+  sv
 }
 
 
@@ -174,7 +158,8 @@ iCheckLinfK <- function(sLinf,sK,type,len) {
     warning(msg,call.=FALSE)    
   } 
   if (sK<0) {
-    if (type %in% c("Typical","typical","Original","original","BevertonHolt","vonBertalanffy","GQ","GallucciQuinn","Schnute")) {
+    if (type %in% c("Typical","typical","Original","original","BevertonHolt",
+                    "vonBertalanffy","GQ","GallucciQuinn","Schnute")) {
       msg <- "The suggested starting value for K is negative, "
     } else {
       msg <- "One  suggested starting value is based on a negative K, "
@@ -338,19 +323,17 @@ iVBStarts.Somers2 <- function(age,len,type,meth0) {
 #=============================================================
 # Static plot of starting values
 #=============================================================
-iVBStartsPlot <- function(age,len,type,sv,ages2use,col.mdl,lwd.mdl,lty.mdl,cex.main) {
+iVBStartsPlot <- function(age,len,type,sv,ages2use,col.mdl,lwd.mdl,lty.mdl,cex.main,col.main) {
   ## attempting to get by bindings warning in RCMD CHECK
   x <- NULL
-  ## Make a title
-  ttl1 <- paste0("von B (",type,")")
-  ttl2 <- paste(paste(names(sv),formatC(unlist(sv),format="f",digits=2),sep="="),collapse=", ")
-  ttl <- paste(ttl1,"with",ttl2)
   ## Plot the data
   # create a transparency value that is the max of 1/2th of the
   # maximum number at any length and age combination or 0.1
   clr <- grDevices::rgb(0,0,0,max(2/max(table(age,len)),0.1))
   # Make the base plot
-  graphics::plot(age,len,pch=19,col=clr,xlab="Age",ylab="Length",main=ttl,cex.main=cex.main)
+  graphics::plot(age,len,pch=19,col=clr,xlab="Age",ylab="Length",
+                 main=paste0("von B (",type,") STARTING VALUES"),
+                 cex.main=cex.main,col.main="red")
   ## Plot the model
   mdl <- vbFuns(type)
   min.age <- min(age,na.rm=TRUE)
@@ -363,139 +346,6 @@ iVBStartsPlot <- function(age,len,type,sv,ages2use,col.mdl,lwd.mdl,lty.mdl,cex.m
     graphics::curve(mdl(x,unlist(sv),t1=ages2use),from=min.age,to=max.age,
                     col=col.mdl,lwd=lwd.mdl,lty=lty.mdl,add=TRUE)
   }
-}
-
-
-#=============================================================
-# Dynamics plots for finding starting values -- main function
-#=============================================================
-iVBStartsDynPlot <- function(age,len,type,sv,ages2use) {
-  ## internal refresh function for the dialog box
-  refresh <- function(...) {
-    p1 <- relax::slider(no=1)
-    p2 <- relax::slider(no=2)
-    p3 <- relax::slider(no=3)
-    iVBDynPlot(age,len,type,p1,p2,p3,ages2use)
-  } # end internal refresh
-
-  ## internal function to make minimum values for the sliders
-  iMake.slMins <- function(sv) {
-    svnms <- names(sv)
-    tmp <- c(Linf=NA,K=NA,t0=NA,L0=NA,omega=NA,t50=NA,L1=NA,L2=NA,L3=NA)
-    if ("Linf" %in% svnms) tmp["Linf"] <- 0.5*sv[["Linf"]]
-    if ("K" %in% svnms) tmp["K"] <- 0.01
-    if ("t0" %in% svnms) tmp["t0"] <- -5
-    if ("L0" %in% svnms) tmp["L0"] <- 0
-    if ("omega" %in% svnms) tmp["omega"] <- 0.5*sv[["omega"]]
-    if ("t50" %in% svnms) tmp["t50"] <- 0.1*sv[["t50"]]
-    if ("L1" %in% svnms) tmp["L1"] <- 0.5*sv[["L1"]]
-    if ("L2" %in% svnms) tmp["L2"] <- 0.5*(sv[["L1"]]+sv[["L2"]])
-    if ("L3" %in% svnms) tmp["L3"] <- ifelse("L2" %in% svnms,0.5*(sv[["L2"]]+sv[["L3"]]),0.5*(sv[["L1"]]+sv[["L3"]]))
-    # reduce to only those in sv
-    tmp <- tmp[which(names(tmp) %in% svnms)]
-    # make sure they are in the same order as in sv
-    tmp[svnms]
-  }  # end iMake.slMins
-  
-  ## internal function to make maximum values for the sliders
-  iMake.slMaxs <- function(sv,age) {
-    svnms <- names(sv)
-    tmp <- c(Linf=NA,K=NA,t0=NA,L0=NA,omega=NA,t50=NA,L1=NA,L2=NA,L3=NA)
-    if ("Linf" %in% svnms) tmp["Linf"] <- 1.5*sv[["Linf"]]
-    if ("K" %in% svnms) tmp["K"] <- 2*sv[["K"]]
-    if ("t0" %in% svnms) tmp["t0"] <- 5
-    if ("L0" %in% svnms) tmp["L0"] <- 2*sv[["L0"]]
-    if ("omega" %in% svnms) tmp["omega"] <- 2*sv[["omega"]]
-    if ("t50" %in% svnms) tmp["t50"] <- 0.6*max(age,na.rm=TRUE)
-    if ("L1" %in% svnms) tmp["L1"] <- ifelse("L2" %in% svnms,0.5*(sv[["L1"]]+sv[["L2"]]),
-                                             0.5*(sv[["L1"]]+sv[["L3"]]))
-    if ("L2" %in% svnms) tmp["L2"] <- 0.5*(sv[["L2"]]+sv[["L3"]])
-    if ("L3" %in% svnms) tmp["L3"] <- 1.5*sv[["L3"]]
-    # reduce to only those in sv
-    tmp <- tmp[which(names(tmp) %in% svnms)]
-    # make sure they are in the same order as in sv
-    tmp[svnms]
-  } # end iMake.slMaxs
-  
-  ## internal function to make delta values for the sliders
-  iMake.slDeltas <- function(sv) {
-    svnms <- names(sv)
-    tmp <- c(Linf=NA,K=NA,t0=NA,L0=NA,omega=NA,t50=NA,L1=NA,L2=NA,L3=NA)
-    if ("Linf" %in% svnms) tmp["Linf"] <- 0.01*sv[["Linf"]]
-    if ("K" %in% svnms) tmp["K"] <- 0.01
-    if ("t0" %in% svnms) tmp["t0"] <- 0.01
-    if ("L0" %in% svnms) tmp["L0"] <- 0.01*sv[["L0"]]
-    if ("omega" %in% svnms) tmp["omega"] <- 0.01*sv[["omega"]]
-    if ("t50" %in% svnms) tmp["t50"] <- 0.01
-    if ("L1" %in% svnms) tmp["L1"] <- 0.01*sv[["L1"]]
-    if ("L2" %in% svnms) tmp["L2"] <- 0.01*sv[["L2"]]
-    if ("L3" %in% svnms) tmp["L3"] <- 0.01*sv[["L3"]]
-    # reduce to only those in sv
-    tmp <- tmp[which(names(tmp) %in% svnms)]
-    # make sure they are in the same order as in sv
-    tmp[svnms]
-  } # end iMake.slDeltas
-  
-  ## Main function
-  ## The default values for the sliders will be at the starting
-  ## values as determined above.  Unlist first to make as a vector.
-  sl.defaults <- unlist(sv)
-  ## Grab names from the sv vector
-  sl.names <- names(sl.defaults)
-  ## Make minimum, maximum and delta values
-  sl.mins <- iMake.slMins(sl.defaults)
-  sl.maxs <- iMake.slMaxs(sl.defaults,age)
-  sl.deltas <- iMake.slDeltas(sl.defaults)
-  ## Make a title
-  sl.ttl <- paste0("Von Bertalanffy (",type,")")
-  ## Set up names that are specific to type and param
-  relax::gslider(refresh,prompt=TRUE,hscale=1.5,pos.of.panel="left",
-                 title=sl.ttl,sl.names=sl.names,
-                 sl.mins=sl.mins,sl.maxs=sl.maxs,
-                 sl.deltas=sl.deltas,sl.defaults=sl.defaults)
-}
-
-#=============================================================
-# Constructs the actual plot in the dynamics plots for finding
-# starting values
-#=============================================================
-iVBDynPlot <- function(age,len,type,p1,p2,p3,ages2use) {
-  ## create a sequence of age values
-  max.age <- max(age,na.rm=TRUE)
-  x <- seq(0,max.age,length.out=20*max.age)
-  ## create a sequence of matching recruit values according to
-  ## the model type and param
-  y <- iVBDynPlot_makeL(x,type,p1,p2,p3,ages2use)
-  ## Construct the scatterplot with superimposed model
-  opar <- graphics::par(mar=c(3.5,3.5,1.25,1.25), mgp=c(2,0.4,0), tcl=-0.2, pch=19)
-  graphics::plot(age,len,xlab="Age",ylab="Mean Length")
-  graphics::lines(x,y,lwd=2,lty=1,col="blue")
-  graphics::par(opar)
-}
-
-#=============================================================
-# Construct values for length given values of age, a Von B
-# model and values of the parameters.  This is called by iVBDynPlot()
-#=============================================================
-iVBDynPlot_makeL <- function(x,type,p1,p2,p3,ages2use){
-  switch(type,
-         Typical=            { # p1=Linf, p2=K,  p3=to
-                               y <- p1*(1-exp(-p2*(x-p3))) },
-         Original=           { # p1=Linf, p2=L0, p3=K
-                               y <- (p1-(p1-p2)*exp(-p3*x)) },
-         GQ=, GallucciQuinn= { # p1=omega,p2=K,  p3=t0
-                               y <- (p1/p2)*(1-exp(-p2*(x-p3))) },
-         Weisberg=           { # p1=Linf, p2=t50,  p3=to
-                               y <- p1*(1-exp(-(log(2)/(p2-p3))*(x-p3))) },
-         Mooij=              { # p1=Linf, p2=L0, p3=omega
-                               y <- p1-(p1-p2)*exp(-(p3/p1)*x) },
-         Francis=            { # p1=L1, p2=L2, p3=L3
-                               r <- (p3-p2)/(p2-p1)
-                               t <- c(ages2use[1],mean(ages2use),ages2use[2])
-                               y <- p1+(p3-p1)*((1-r^(2*((x-t[1])/(t[3]-t[1]))))/(1-r^2)) },
-         Schnute=            { # p1=L1, p2=L3, p3=K
-                               t <- ages2use
-                               y <- p1+(p2-p1)*((1-exp(-p3*(x-t[1])))/(1-exp(-p3*(t[2]-t[1])))) }
-  ) # end type switch
-  y
+  ## Put the starting values to put on the plot
+  legend("bottomright",paste(names(sv),formatC(unlist(sv),format="f",digits=2),sep="="),bty="n")
 }